requirements.txt
pycryptodome
huffman
PyWavelets
pillow
numpy
scikit-learn
scikit-image
